requirements.txt
statsmodels==0.12.1
lightgbm==3.3.2
scikit-learn==0.24.2
numpy==1.16.4
pandas==1.1.4
ray==1.0.1.post1
scipy==1.5.4
tensorflow-cpu==2.3.1
keras==2.4.3
fasttreeshap==0.1.1
numba==0.53.1